Fix cache issue causing imaging to fail
Problem:
UnboundLocalError: local variable properties referenced before assignment

Solution:
Fix the underlying issue, that the parent image was None

When an instance is imaged, a snapshot is made of the instance, then that
snapshot is mounted locally so that the filesystem can be cleaned. Then that
cleaned image is then uploaded as the image, which will appear in atmosphere.

When chromogenic images, it actually checks if a snapshot has been made before
and just uses that rather than always creating a snapshot.

Just prior to uploading chromogenic looks up the snapshot image (what it calls
the parent image), and uses properties on it for the final image. The bug was
that looking up the snapshot would return None sometimes, which would cause
the properties lookup to fail.

Chromogenic would first populate the cache, when it checked if the snapshot
had ever been created. Then it would create the snapshot (never updating the
cache) then would later ask if the cache had the snapshot.

So the error can be reliably reproduced, if you delete the snapshot before you
retry imaging.